@@ -1,11 +1,12 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
+import ReactDOM from 'react-dom/client'
 import reportWebVitals from './reportWebVitals'
-import { App } from './components/App'
+import App from 'components/App'
 
-ReactDOM.render(
-  <App />,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(document.getElementById('root'))
+
+root.render(
+  <App />
 )
 
 // If you want to start measuring performance in your app, pass a function
