@@ -1,28 +1,35 @@
 import React from 'react'
+import { Provider } from 'react-redux'
+import configureStore from '../configureStore'
 import {
-  BrowserRouter as Router,
-  Switch,
-  Route,
+  createBrowserRouter,
+  RouterProvider
 } from 'react-router-dom'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import './App.css'
 import { Game } from './Game'
 import { Home } from './Home'
 
-export const App = () => {
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Home />
+  },
+  {
+    path: '/gitclicker',
+    element: <Game />
+  }
+])
+
+const store = configureStore()
+
+const App = () => {
   return (
-    <>
+    <Provider store={ store }>
       <CssBaseline />
-      <Router>
-        <Switch>
-          <Route exact path="/">
-            <Home />
-          </Route>
-          <Route path="/">
-            <Game />
-          </Route>
-        </Switch>
-      </Router>
-    </>
+      <RouterProvider router={router} />
+    </Provider>
   )
 }
+
+export default App
